Move streams views into their own file.
