+++ zproject/urls.py
@@ -106,21 +106,21 @@ urlpatterns += patterns('zerver.views',
     # These are json format views used by the web client.  They require a logged in browser.
     url(r'^json/update_pointer$',           'json_update_pointer'),
     url(r'^json/get_old_messages$',         'messages.json_get_old_messages'),
-    url(r'^json/get_public_streams$',       'json_get_public_streams'),
-    url(r'^json/rename_stream$',            'json_rename_stream'),
-    url(r'^json/make_stream_public$',       'json_make_stream_public'),
-    url(r'^json/make_stream_private$',      'json_make_stream_private'),
+    url(r'^json/get_public_streams$',       'streams.json_get_public_streams'),
+    url(r'^json/rename_stream$',            'streams.json_rename_stream'),
+    url(r'^json/make_stream_public$',       'streams.json_make_stream_public'),
+    url(r'^json/make_stream_private$',      'streams.json_make_stream_private'),
     url(r'^json/send_message$',             'messages.json_send_message'),
     url(r'^json/invite_users$',             'json_invite_users'),
     url(r'^json/bulk_invite_users$',        'json_bulk_invite_users'),
     url(r'^json/settings/change$',          'json_change_settings'),
     url(r'^json/notify_settings/change$',   'json_change_notify_settings'),
     url(r'^json/ui_settings/change$',       'json_change_ui_settings'),
-    url(r'^json/subscriptions/remove$',     'json_remove_subscriptions'),
-    url(r'^json/subscriptions/add$',        'json_add_subscriptions'),
-    url(r'^json/subscriptions/exists$',     'json_stream_exists'),
-    url(r'^json/subscriptions/property$',   'json_subscription_property'),
-    url(r'^json/get_subscribers$',          'json_get_subscribers'),
+    url(r'^json/subscriptions/remove$',     'streams.json_remove_subscriptions'),
+    url(r'^json/subscriptions/add$',        'streams.json_add_subscriptions'),
+    url(r'^json/subscriptions/exists$',     'streams.json_stream_exists'),
+    url(r'^json/subscriptions/property$',   'streams.json_subscription_property'),
+    url(r'^json/get_subscribers$',          'streams.json_get_subscribers'),
     url(r'^json/fetch_api_key$',            'json_fetch_api_key'),
     url(r'^json/update_active_status$',     'json_update_active_status'),
     url(r'^json/get_active_statuses$',      'json_get_active_statuses'),
@@ -176,16 +176,6 @@ urlpatterns += patterns('zerver.views',
 v1_api_and_json_patterns = patterns('zerver.views',
     url(r'^export$', 'rest_dispatch',
             {'GET':  'export'}),
-    url(r'^streams$', 'rest_dispatch',
-            {'GET':  'get_streams_backend'}),
-    # GET returns "stream info" (undefined currently?), HEAD returns whether stream exists (200 or 404)
-    url(r'^streams/(?P<stream_name>.*)/members$', 'rest_dispatch',
-            {'GET': 'get_subscribers_backend'}),
-    url(r'^streams/(?P<stream_name>.*)$', 'rest_dispatch',
-            {'HEAD': 'stream_exists_backend',
-             'GET': 'stream_exists_backend',
-             'PATCH': 'update_stream_backend',
-             'DELETE': 'deactivate_stream_backend'}),
     url(r'^users$', 'rest_dispatch',
             {'GET': 'get_members_backend',
              'POST': 'create_user_backend'}),
@@ -196,19 +186,11 @@ v1_api_and_json_patterns = patterns('zerver.views',
     url(r'^users/me/pointer$', 'rest_dispatch',
             {'GET': 'get_pointer_backend',
              'PUT': 'update_pointer_backend'}),
-    # GET lists your streams, POST bulk adds, PATCH bulk modifies/removes
-    url(r'^users/me/subscriptions$', 'rest_dispatch',
-            {'GET': 'list_subscriptions_backend',
-             'POST': 'add_subscriptions_backend',
-             'PATCH': 'update_subscriptions_backend'}),
     url(r'^users/me/alert_words$', 'rest_dispatch',
             {'GET': 'list_alert_words',
              'PUT': 'set_alert_words',
              'PATCH': 'add_alert_words',
              'DELETE': 'remove_alert_words'}),
-    url(r'^default_streams$', 'rest_dispatch',
-            {'PATCH': 'add_default_stream',
-             'DELETE': 'remove_default_stream'}),
     url(r'^realm$', 'rest_dispatch',
             {'PATCH': 'update_realm'}),
     url(r'^users/me/api_key/regenerate$', 'rest_dispatch',
@@ -225,7 +207,7 @@ v1_api_and_json_patterns = patterns('zerver.views',
          'DELETE': 'remove_android_reg_id'}),
     url(r'^users/(?P<email>.*)/reactivate$', 'rest_dispatch',
             {'POST': 'reactivate_user_backend'}),
-    url(r'^users/(?P<email>.*)$', 'rest_dispatch',
+    url(r'^users/(?P<email>[^/]*)$', 'rest_dispatch',
             {'PATCH': 'update_user_backend',
              'DELETE': 'deactivate_user_backend'}),
     url(r'^bots$', 'rest_dispatch',
@@ -252,6 +234,26 @@ v1_api_and_json_patterns = patterns('zerver.views',
             {'GET':  'render_message_backend'}),
     url(r'^messages/flags$', 'rest_dispatch',
             {'POST':  'update_message_flags'}),
+
+) + patterns('zerver.views.streams',
+    url(r'^streams$', 'rest_dispatch',
+        {'GET':  'get_streams_backend'}),
+    # GET returns "stream info" (undefined currently?), HEAD returns whether stream exists (200 or 404)
+    url(r'^streams/(?P<stream_name>.*)/members$', 'rest_dispatch',
+        {'GET': 'get_subscribers_backend'}),
+    url(r'^streams/(?P<stream_name>.*)$', 'rest_dispatch',
+        {'HEAD': 'stream_exists_backend',
+         'GET': 'stream_exists_backend',
+         'PATCH': 'update_stream_backend',
+         'DELETE': 'deactivate_stream_backend'}),
+    url(r'^default_streams$', 'rest_dispatch',
+        {'PATCH': 'add_default_stream',
+         'DELETE': 'remove_default_stream'}),
+    # GET lists your streams, POST bulk adds, PATCH bulk modifies/removes
+    url(r'^users/me/subscriptions$', 'rest_dispatch',
+        {'GET': 'list_subscriptions_backend',
+         'POST': 'add_subscriptions_backend',
+         'PATCH': 'update_subscriptions_backend'}),
 
 ) + patterns('zerver.tornadoviews',
     url(r'^events$', 'rest_dispatch',
